Added code to trap emulation crash and save console output file
Whenever tzar emulation was crashing, it would not close the console
sink file and the important information about what went wrong would be
lost, particularly information about random seed settings.
This new code wraps the main call in a tryCatch() and makes sure that
the console output file is saved when there is a fatal error in a run.

Tested by adding these 3 lines at the start of bdgpxupaper::xu_paper_main()

cat ("\n\nAbout to crash tzar to test saving the console sink.\n")
stop ("Test crash of tzar and bdpgxupaper")
cat ("\n\nImmediately after crash tzar to test saving the console sink.  SHOULD NOT REACH THIS STATEMENT.\n")

The corresponding commit in bdpgxupaper for the test was:
- 0a297248
- Special version to test new tzar trapping fatal errors & save console output
- 2017-12-13 08:30 (strangely not Melbourne time...)

+++ R/mainline_support_functions.R
@@ -170,13 +170,64 @@ clean_up_after_tzar_emulation <- function (tzar_in_progress_dir_name,
     cat ("\n\nFinal tzar output is in:\n    '", tzar_emulation_completed_dir_name,
          "'\n\n", sep='')
 
-    file.remove (emulation_scratch_file_path)
+    if (file.exists (emulation_scratch_file_path))
+        {
+        file.remove (emulation_scratch_file_path)
+        }
 
     clean_up_console_sink (echo_console_to_temp_file,
                            console_sink_file_info,
                            full_output_dir_with_slash)
 
     file.rename (tzar_in_progress_dir_name, tzar_emulation_completed_dir_name)
+     }
+
+#===============================================================================
+
+handle_error_clean_up_after_tzar_emulation_crash <-
+                                        function (tzar_in_progress_dir_name,
+                                                  emulation_scratch_file_path,
+                                                  echo_console_to_temp_file,
+                                                  console_sink_file_info,
+                                                  full_output_dir_with_slash)
+    {
+        #----------------------------------------------------------------
+        #  Write to output to flag that the emulation run crashed.
+        #  Using message() to make it bright red and hard to miss in
+        #  console output.
+        #  However, that message doesn't seem to show up in the console
+        #  sink file, so write the same message out using cat too.
+        #----------------------------------------------------------------
+
+    cat ("\n**********  TZAR EMULATION RUN CRASHED  **********",
+         "\n**********       CLEANING UP NOW        **********\n")
+
+    message (paste0 ("\n**************************************************",
+                     "\n**********  TZAR EMULATION RUN CRASHED  **********",
+                     "\n**********       CLEANING UP NOW        **********",
+                     "\n**************************************************"))
+
+        #----------------------------------------------------------------
+        #  Remove the emulation scratch file and move the console sink
+        #  file to the metadata area of the current run that is being
+        #  aborted.
+        #----------------------------------------------------------------
+
+    cat ("\nCurrent tzar output area is:\n    '", tzar_in_progress_dir_name,
+         "'", sep='')
+
+    if (file.exists (emulation_scratch_file_path))
+        {
+        cat ("\nRemoving emulation scratch file = '", emulation_scratch_file_path, sep='')
+        file.remove (emulation_scratch_file_path)
+        }
+
+    if (echo_console_to_temp_file)
+        {
+        clean_up_console_sink (echo_console_to_temp_file,
+                               console_sink_file_info,
+                               full_output_dir_with_slash)
+        }
      }
 
 #===============================================================================
